Small refactor.
- Separate out nested classes in non-obvious locations.
- Move spec files around so it's a little more clear what they're
  representing.
