+++ spec/alephant/broker/component_meta_spec.rb
@@ -1,4 +1,4 @@
-require_relative "spec_helper"
+require "spec_helper"
 
 describe Alephant::Broker::ComponentMeta do
   let(:id) { "foo" }
