+++ spec/alephant/broker/load_strategy/http_spec.rb
