+++ spec/alephant/broker/load_strategy/s3/archived_spec.rb
