+++ spec/spec_helper.rb
@@ -8,6 +8,8 @@ end
 
 require "pry"
 require "json"
+require "rack/test"
+
 require "alephant/broker"
 require "alephant/broker/load_strategy/s3/sequenced"
 require "alephant/broker/load_strategy/s3/archived"
@@ -15,7 +17,6 @@ require "alephant/broker/load_strategy/http"
 require "alephant/broker/cache"
 require "alephant/broker/errors/content_not_found"
 
-require "rack/test"
 
 ENV["RACK_ENV"] = "test"
